Move classes to Impl namespace
This reveals that ConsumerWorkService should not be public, but it's
tricky to make that happen.
