+++ projects/client/RabbitMQ.Client/src/client/impl/ConsumerWorkService.cs
@@ -3,7 +3,7 @@ using System.Collections.Concurrent;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace RabbitMQ.Client
+namespace RabbitMQ.Client.Impl
 {
     public class ConsumerWorkService
     {
